requirements.txt
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
weasyprint>=60.0
